don't wrap a function around the declarations of class members
I think the use of a function here dates back to when we used to
emit the typeAnnotationsHelper as a method on the class.  And we
did that because we needed the resulting code to TS type-check.

None of that is true anymore, and Closure style is to declare members
etc. at the top level, not in a helper.