Skip emitting clutz aliases for default exports.
The current emit for default exports is broken and the alternative path from tsickle to clutz is also broken, so no one is currently relying on the current behaviour.  Since the style guide forbids default exports, the easiest way to get error free .d.ts files is to skip emitting aliases for default exports.