Fix and add tests for googmodule comment emitting.
Previously, when tsickle ran only the googmodule transformer, e.g. when
both type annotation and decorator conversion is false, file level
comments would not be emitted, causing problems with some load bearing
comments.

googmodule in practice runs in two quite different modes:
1. after tsickle source map transformers have run
2. standalone (e.g. when running plain TS and only doing googmodule)

In (1), tsickle runs various transforms (including some nasty hacks)
beforehand, synthesizing comments, setting text ranges on some nodes,
and so on, to make things work with the source map based transform.

In (2), no such thing happens, and so googmodule transformer has to
operate on a vanilla TS AST.

This change adds tests for both situations, and in both googmodule_test
and in the end to end tests, with the .puretransform monicker.

The actual fix for the problem with not emitting comments in pure
transformation mode then was that googmodule did not set the text range
on the new statements in the transformed source file. This caused
TypeScript to skip emit of the leading file comment ranges.