Do not export the function using a global this.
It turns out JSCompiler has a second warning for using `this` in a
static function. The warning is triggered if the function is exported,
because it gets inlined into the exports object and is then considered
to be a static function. This warning cannot be disabled and isn't
included in the `globalThis` suppression, so this fails in newer Closure
Compiler versions.

The fix is to simply not export the function - we're only testing that
the comment gets retained here, so whether or not the function is
exported is irrelevant. This still adds an empty `export {}` block so
that your IDE doesn't complain about duplicate declarations of `f`.