Emit exported typedefs for use in the local scope.
Previously, tsickle would emit an `exports.Foo` typedef, but then
references to this typedef in the local scope would use just `Foo`,
which breaks compilation.

Closure however does not support referencing `exports.Foo` within a
module. The fix is to emit both a local variable for the type alias, and
an export aliasing the type alias (`exports.Foo = Foo;`).