Support property access on imported types.
TypeScript uses forward declares to refer to imported modules in type
annotations to match TS' loading semantics:

    import {SomeImport} from 'some-module';
    const foo: SomeImport = ...;

tsickle will introduce a forward declare for the imported module:

    var tsickle_fwd_decl_1 = goog.forwardDeclare('some-module')'
    const /** @type {!tsickle_fwd_decl_1.SomeImport} */ foo = ...;

TypeScript code can use types nested in an imported symbol:

    import {SomeImport} from 'some-module';
    const bar: SomeImport.SomeNestedType = ...;

Previously, tsickle would only use an alias for a top level symbol
directly (`SomeImport`), but not for a symbol accessed on an aliased
symbol. This change has tsickle consult the alias map for all symbols
encountered in a type expression.

    const bar: SomeImport.SomeNestedType = ...;
    // gets emitted as:
    const /** @type {!tsickle_fwd_decl_1.SomeImport.SomeNestedType} */ foo = ...;

This also required a change to make sure tsickle does not introduce
aliases for exported declared namespaces, as those do not have an
exported symbol that can be aliased but should rather fall back to the
global symbol emitted for them (covered by existing tests).