Add a test for non-null generics.
In Closure Compiler, generic types (`T`) are always nullable, with no
straight forward syntax to express a non-nullable generic type (*).
Normally this code would produce a warning, but the `@suppress
{checkTypes}` prevents that successfully.

This change is a pure regression test to make sure this stays the same.

(*) Closure Compiler's Type Transformation Language (TTL) allows
expressing the concept, but leads to very cumbersome types:

    /**
    * @template T
    * @param {(null|T)} value
    * @param {!TNotNull} def
    * @return {!TNotNull}
    * @template TNotNull :=
    *     cond(isUnknown(T), unknown(),
    *       mapunion(T, (X) =>
    *         cond(eq(X, 'null'), none(), X)))
    * =:
    */
    function getOrDefault(value, def) {
        return def;
    }

Given the error is successfully repressed, this doesn't seem worth
implementing in tsickle.