Fix comment emit inside for arrow functions.
Previously, tsickle would synthesize a comment for expression bodies of
arrow functions:

    var x = () =>
        // comment
        1;

This would cause the comment to be emitted in between the return
statement and the expression after TS transform:

    var x = function() {
        // comment
        return
        // comment
        1;
    }

The first comment is TS correctly transforming the arrow function, the
second comment is tsickle re-emitting the synthesized comment, causing
Automatic Semicolon Insertion and thus breaking the code.

The fix is to just not do that - TypeScript emits comments in the
expression correctly, and tsickle already avoids modifying the comment
exactly because emitting comments in arrow functions is dangerous.

Fixes #802.