Export enums declared in namespaces.
tsickle normally splits enum declaration and export, so that just `Enum`
in the local scope of the file resolves to the correct symbol:

    /** @enum {...} */
    const Enum = { ... };
    export {Enum};

This works for ES6 exports, but TypeScript's namespaces do not support
separate export statements. To avoid the problem, disable processing of
exported enums that are declared in namespaces.

This fixes the runtime behaviour (`x.Enum` resolves to the right value).
It also means enums in namespaces no longer introduce Closure types.
However tsickle's emit for namespaces is not understood by Closure in
either case, so this fixes runtime behaviour while at least not
regressing otherwise.

This also required resetting the text range on module declarations to
prevent duplicate comment emits on them. While at it, this change
removes some outdated workarounds for long fixed TypeScript issues.

Fixes #805.