Run the full tslint suite in npm test.
tslint has two sets of linter rules, the purely syntactic ones, and the
ones that require type information, which require a ts.Program.
gulp-tslint allegedly has support for running type checked rules, but
in practice this seems to be broken.

Running tslint from the command line works, is simpler anyway, and
reduces our overall dependencies. So this change removes the gulp-tslint
dependency in favour of just calling plain tslint from npm test.

This change also fixes a large number of unused imports that have gone
undetected for a long time due to not running these rules.