Add web animations typings to closure_externs.js.
The typings for web animations were just added to the typescript standard library.  They partially replace the DefinitelyTyped typings for web animations.  The closure platform externs don't include web animations, which was previously fine, since tsickle produced the externs from the DefinitelyTyped typings.  With 2.8, many of the typings are in the standard library, so closure_externs.js needs to contain the corresponding externs.