Fix the definition of SymbolConstructor by relaxing it.
The definition is incorrect, `Symbol` is not invoked using `new`, but
rather just as a function. A better approximation of the type would be
`function(string):Symbol`.

Closure Compiler has a builtin for `symbol` and a definition of
`Symbol`, but no type or construct to express `typeof Symbol`.

Rather than trying a partial emulation here, prefer using the unknown
type (`?`), which is likely to give fewer problems.