drop use of the google-closure-compiler npm module
This module depends on a bunch of other npm modules (like vinyl
and so on) that we don't use, which makes it difficult to run
inside of Google.

Instead, rewrite just the tiny bit of the compiler API that we want.

Note: this keeps the google-closure-compiler npm dep in package.json
just for access to the compiler .jar file.  Users can override which
jar to use by setting compiler.jarPath.