drop chai and use jasmine expect instead
When we switched from mocha to jasmine, we kept our dependency on
chai, but this is confusing because jasmine expects you to call its
expect() at least once.  Also, chai has these cutesy testers that
have side effects on property get that make me and tslint sad.

This change drops chai, brings in jasmine-diff (for diff output),
and switches all the test assertions to use the jasmine equivalent.