Export types as non-nullable `@typedefs`.
In Closure Compiler's `@typedef`, any mentioned types without a `!` are
implicitly nullable (as in all other locations). While `typeToClosure`
takes care to insert the `!` as needed, some locations in tsickle export
a plain type name. In that case, we must take care to insert the `!`
at the emit site.

This fixes a crash in Closure Compiler when as it encounters a union
type when resolving super types (e.g. implemented interfaces).