use tslib and clean up use of TypeScript helpers
TypeScript used to have 'helpers' for builtins like __decorate,
which tsickle had special cases for.  The helpers eventually
moved to the 'tslib' npm module.

Clean up tsickle to match the new tslib world.
We accidentally had two copies of the helpers; delete them both,
and check in the current version of our tslib modification.

This additionally required some hacking in the BUILD files to make
tslib visible to all the test suites.