Handle Bundle objects in googmodule transformer.
TS2.9 will pass Bundle objects through post-module transformers if the
`outFile` option is being used. While tsickle generally doesn't support
this option, this change avoids crashing on it in a non-obvious way.

The TS change is not actually expressed in the API, transformers still
receive `ts.SourceFile`, not `ts.Bundle`, so this requires casting
through any.