Reinstate the E2E closure test.
During the last `BUILD` verschlimmbesserung, the `closure_e2e_test`
Jasmine test ended up depending on the `e2e_test_lib`, not
`closure_e2e_test_lib` as intended, causing the Closure test to never
execute.

On the bright side, this executed all other E2E tests twice, for added
safety. Sigh. The moral is that we should not be trusted around test
infrastructure.

This also removes the `duplicated_externs` test, which does not work
with the Closure version used in tsickle.