Fix tests using outFile in tsickle.
tsickle fundamentally cannot handle `outFile`, so these tests were a bit
nonsensical. Others apparently just used `outFile` for convenience. This
change deletes the former and adjusts the latter to use separate source
map files.

The medium term future for this code is to be deleted once tsickle is
fully on transformers.