rearrange @fileoverview handling
The fileoverview logic had three different cases where it added to
the suppress list, and extraRequire was only included in one of the
three.  I want to add another entry to our suppression list, so I'll
start by refactoring this so that suppressions are handled in one place.

This rebaselines ~every test because we now consistently suppress
extraRequire.