use es6 form when generating export statement
This produces equivalent output when targeting goog.module
(because it becomes an exports.foo = foo line) so no tests are
affected.

But when tsickle is emitting ES6, this stays as an export {}
statement, which is the proper output for that.

Note: apparently tsickle has no tests covering the ES6 output
mode(!), or this would've been caught earlier.

Fixes #848.