Avoid typedef name conflicts.
Introducing a new name in a scope can cause several different problems,
such as incorrect references or accidentally duplicated exports.

This change introduces a name mangling scheme for type aliases, so that
the newly introduced value symbols do not (likely) cause conflicts.

This problem in tsickle was exposed, but not caused, by #849. The issue
always existed, it was just suppressed by users and went unnoticed
(partially due to untyped mode). The exports change caused an
unsuppressible error that this change should fix, together with the
previous ones.