Emit hard goog.require()s only for explicitly imported modules.
Previously, tsickle would emit goog.require() calls for any symbol that
was referenced in the current file if its module contained only types,
regardless of whether the module was actually imported or not.

This can cause strict dependency checking errors if the hard required
file was not already a valid dependency of the current file.

This change avoids the problem by being careful not to emit such imports
if the referenced module is not explicitly imported.