drop gulp in favor of our own check_format script
The only reason we have gulp involved is for checking clang-format.
I'd like to also check that README.md is formatted.  To do that
we'll need more formatter logic.  Rather than add more gulp, I figured
I'd go the other way and remove the gulp in this change first.

PS: I wrote the checker as a JS program rather than a shell script to
have some hope of this working on Windows.