Extract externs writing out of tsickle.ts.
This is a standalone version of the externs writing, traversing the AST
and emitting text (as opposed to a TypeScript transformer that converts
to AST nodes). Given the simple format of externs, and that TypeScript's
transformers do not allow adding additional output files, this seems
like the most appropriate solution for externs writing.