Only emit a single leading @fileoverview comment.
Source files can include @fileoverview comments in surprising locations,
e.g. in the middle of module body statements. Because Closure Compiler
ignores all but the last @fileoverview comment, and tsickle writes its
suppressions etc in the first fileoverview comment, tsickle must not
emit any other @fileoverview comments.

This change makes sure to accept multiple leading comments (which
commonly happens with @license and @fileoverview in one file), and then
suppresses emit for any other @fileoverview comments.

The alternative would be to look for any @fileoverview comments strewn
across the source file. That's possible, but complex (what if it's in a
nested block? what if we find multiple?), and probably more confusing
than helpful in the long run.

This is an intermediate state to unblock the new transformer based emit
landing. In a follow up change, the comment suppression will be turned
into an error, and existing code must be cleaned up.