Skip casts and parentheses in extends clauses.
Closure Compiler reports an error for `extends` clauses that do not
contain a simple identifier or qualified (dotted) name. tsickle's move
to transformers improves its emit to retain comments added in extends
clauses, triggering this issue.

The problem is likely an oversight in Closure, but this change works
around until the bug is fixed in Closure. It does so by skipping any
casts or parentheticals in extends clauses.