wrap interface declarations in if(false)
In our upcoming world of transformers, it's easier if we consistently
always wrap interface member declarations in an if(false) block.

This partially rolls back some of 249273f6417ea3696934c5b71abbbf4b202ce612,
which dropped function wrapper in favor of either using if(false) or
no wrapper at all for interfaces, so now we consistently use if(false).