Report warnings for duplicated and run-in file comments.
In TypeScript, file level comments are comments that are separated from
the file body with an empty line. In Closure, file level comments are
comments marked with a particular (often load bearing) JSDoc tag, e.g.
`@fileoverview` or `@pintomodule`.

If a TypeScript contains such a Closure file level comment in a comment
that is *not* a TypeScript file level comment, TypeScript will often not
emit the comment at all, e.g. because it was placed on an interface or
type import that will not be emitted. That violates user expectations,
as their (load bearing) comment is now silently gone.

This change has tsickle report warnings whenever it encounters a file
comment that's either not properly separated from the file body, is
duplicated, or appears after the initial statement in the body of the
file. This will typically break compilation first party code so that the
situation can be discovered and fixed early by the user, and ignored for
third party code (where it's outside the user's control).