consistently use absolute paths
The TypeScript API wants absolute paths, and fights against you if
you don't use them.  This change makes tsickle assert that it has
absolute paths in various places and fixes up many of the places
where we accidentally weren't using absolute paths.

This changes the public API of tsickle in a few places:
- source files coming in to tsickle must be absolute
- other path-based attributes like the based type blacklist and externs
  also work with absolute paths

To implement this, in the whole test suite we depend on the compiler
options always including a rootDir, so the various tests now say
options.rootDir! occasionally.  Also each test's path manipulation
has been adjusted to be careful to use absolute/relative paths where
appropriate.

This change also removes calls to process.cwd() and path.resolve(),
both of which depend on the current directory, in favor of reliably
using bazel's runfiles.