Emit non-parameter template types as non-nullable.
TypeParameter type's symbols are usually actual type parameters (`<T>`).
In some situations, such as bound parameters in the case of a `this`
return type, they can reference regular type symbols, such as a class or
interface.

Type parameters are non-nullable in Closure, where references to other
symbols are nullable. To avoid emitting a nullable type where a
non-nullable is meant, insert the `!` non-nullable operator for
non-type-parameter symbols.