add a flag for goog.requireType instead of goog.forwardDeclare
goog.requireType more closely matches the semantics of a TypeScript
type import.  This change adds the support for goog.requireType,
but keeps it off behind a flag while we work through some remaining
issues.

As part of this change, clean up a few of the comments in the area
and simplify the code a bit.