consistently avoid emitting inexpressible heritage clauses
There are many TypeScript patterns around heritage (implements/extends)
that are inexpressible in Closure.  We were inconsistent about what
we did in those situations.

This change introduces a type named `InexpressibleType` that we emit in
any place where things don't work.  This serves as a hint to humans
looking at the typed output that we dropped something.

Finally, this test adds a test case for
  class X extends Partial<Other>
which is also inexpressible in Closure, and fixes our previously wrong
emit for that output.