Adjust tsickle to getCombinedModifierFlags change.
`getCombinedModifierFlags` now takes a `ts.Declaration`. tsickle has a
number of locations calling it on arbtirary nodes or statements (through
the helper `hasModifierFlags`).

The function still works on both, so with the casts the code works.
Adjusting our helper and pushing the casts out to the usage sites will
allow us to migrate the code to be type safe over time.