restore 'implements class' hack
tsickle previously converted
  class X implements OtherClass
into
  class X extends OtherClass
in some situations.

In e466dc5f I removed this behavior, but it appears we have code that
actually relies on this.  I now think this behavior was a mistake but for
now let's preserve our behavior so that we can get the other benefits of
this change.