don't dereference aliases in heritage clauses
If A implements B, we should refer to the B symbol directly, rather than
attempting to resolve B to some other symbol, because that other symbol
may refer to a module that hasn't been imported.

This change makes us more careful around symbol handling in heritage
clauses, as a model for how we should handle symbols in general: there's
a "local" symbol, which is the file-local, imported name that should be
used in type annotations, and then separately there's the "true" symbol,
which is the symbol after following aliases, which has info like whether
it refers to a type or a value.