drop defn and emit of InexpressibleType
When tsickle cannot express the implements/extends of a type, it
would emit some
  @implements {InexpressibleType}
as a hint to humans.

I now see this was a bad idea, because it is visible to Closure
compiler and fails in situations where e.g. a class has multiple
of these.  Additionally, we shouldn't be dumping new Closure types
like InexpressibleType into the global namespace in the first place.

I have another change that changes our emit for inexpressible types,
but this smaller change first just deletes this unneeded comment,
because even this comment itself causes some build failures.