Handle single-value enums.
For enums that only have a single value, TypeScript's
getBaseTypeOfLiteralType returns the EnumLiteral type for that value
instead of the type for the enum. Previously, this caused tsickle to
emit a type name representing the single enum value, not the enum type,
which is illegal in Closure.

This change works around the problem by first trying the proper API, and
then falling back to using the `symbol.parent` if we still get an
`EnumLiteral` type returned.

See also https://github.com/Microsoft/TypeScript/issues/28869.