Handle rest parameters with template constraints.
Rest parameters must be array types in TypeScript. tsickle unwraps them
to emit the underlying type for Closure. However the Array type in
TypeScript may be expressed in a template parameter constraints, instead
of the parameter itself:

    function f<T extends SomeType[]>(... ts: T) {}

Handle this specific case by resolving the template parameter
constraint, then emitting the underlying type nested in it.

Fixes #944.