Blacklist based on a symbol's declaration path.
When emitting forward declarations, tsickle would previously inspect the
import path given at the use site. However the list of blacklisted
locations is a set of (absolute) paths of files that contain symbols
that must not have types emitted for them. These paths can but do not
have to be related to import paths - for example for ambient external
modules (`declare module 'ambient' {...}`), the import path has no
relationship to the declaration site's path.

The fix is to check the declaration site forof the import's module
symbol, which gives the correct path.