Emit types for arrow/function expressions.
Previously, tsickle did not emit types for arrow functions and function
expressions, mostly because it was too hard to correctly pipe the
comments through the parse-emit-parse again pipeline, before tsickle was
a proper transformer.

This change adds appropriate comments, and makes sure to parenthesize
the function expressions, to avoid triggering automatic semicolon
insertion.