carefully handle @template in jsdoc handling
Some @tags are disallowed in the user's *input* jsdoc, like
@private, because we don't want users writing them.

Other @tags are escaped in the tsickle *output* jsdoc, because
we don't want Closure compiler interpreting them.

It turns out @template is special in that we want to allow it
in the input, and we want tsickle to be able to generate it
in the output, but we don't want the user's input to make it
to the output.

See the test case for some more rationale.