Emit inline symbols in `.d.ts` without `declare`.
tsickle tries to carefully not emit non-ambient types in externs. To do
so, it looks for the `declare` keyword on type declarations and their
ancestors.

However in `.d.ts` files the `declare` keyword is optional, so tsickle
should also take into account whether the file itself is a declaration
file.