Remove tsickle auto-quoting transformer.
When this worked it worked well, but when it misfired users found it
very confusing (noone expected that it existed) and suprising to work
with.

Ultimately, it made the emit type-senstitive which is not a feature TS
developers expect.

Instead, we have replaced this with a tsetse check
https://tsetse.info/property-renaming-safe. It follows the same
detection rules, but errors instead of silently fixing the code.

I plan to further followup this thread with a proposal for
--strictPropAccess on Microsoft/TypeScript.