Add suppression for constantProperty check in tsickle emit. (#1014)
We are doing this mostly for supporting namespaces. When one declares
namespace foo.bar the emit implicilty reopens foo, which results in a `
|| export.foo = {}` potential reassignment in the emit. See added test
for the actual code.

In Closure export.foo is implicitly const, so this triggers an error.
For now, we just suppress it using @supress, since TS has checked the
const-ness of the original file.