remove direct dependencies on node builtin modules
For 'assert', just replace it with a throw.

For 'path', use the TypeScript's path manipulation functions.  These
are likely better than using node's path functions anyway because we
interact heavily with TypeScript and using their functions ensures we
match.  On the downside, we must access internal functions to do this.

This is a step towards making it possible to run tsickle in a browser
playground.