don't emit type aliases when they conflict with module imports
In code like

```
import * as x from 'y';
type x = ...;
```

We don't want to emit a `@typedef` for `x`, because it conflicts with
the module import.  We already tried to avoid this problem by checking
whether `x` was both a type and a value, but in this case, `x` is
both a type and an *alias* to the module, and the value-ness of `x`
is on the other side of the alias.

It turns out this pattern of checking for value-ness occurs a few times
in tsickle and the code was inconsistent about whether it expanded
through aliases, so I refactored things to one helper.