drop the CompilerHost parameter to emit
tsickle doesn't actually need a CompilerHost.  It needs a
ModuleResolutionHost (a superclass of CompilerHost) but it already takes
that from the TsickleHost parameter.  Callers of this code often end up
passing the same parameter for both TsickleHost and CompilerHost because
they construct one superobject anyway.

Dropping this helps reduce the API surface, but it in particular makes
things better when running tsickle on a web page, where we might have a
language service but no CompilerHost.