make writeFile parameter non-optional
In my recent refactorings of the tsickle.emit function, I removed the
CompilerHost parameter.  The function used to accept an optional
writeFile callback, which would fallback to the CompilerHost.writeFile.
When I removed CompilerHost I made the writeFile usage optional.

However, it is important that we always intercept writeFile, because we
use it to add the clutz dts aliases; if a caller to tsickle.emit doesn't
provide writeFile we would silently drop that.  The fix is to make this
parameter non-optional, which also improves some of the call sites
(which mostly provided it anyway).