Transform down-leveled await expressions.
TypeScript down-levels await expressions to a library call, passing a
generator function with the tail of the async function:
    tslib_1.__awaiter(this, void 0, void 0, function* () { ... });

The generator function containing the remainder of the function's code
needs an `@this` annotation for Closure to understand `this.` accesses
within it.

Additionally, because all async functions unconditionally pass
`tslib_1.__awaiter(this, ...)`, even functions without a `this` context
need to be annotated with `@this {*}` (i.e. an unknown this type) to
avoid a warning by Closure Compiler.