add tsconfig.json for test_files/
Currently no tsconfig covers these files.  In that situation, TypeScript
picks an implied tsconfig for these when you load them in the editor,
which then causes warnings for duplicate symbols because we check in
both the .ts input and the .js output.

If we add these files to the root tsconfig.json, then our CI runs tslint
over test_files, which we don't want.

So instead we add a test_files/tsconfig.json.
By splitting the tsconfigs like this, we can keep the root tsconfig to
cover specifically the settings that are relevant to tsickle, while the
test_files tsconfig can then include just the test-relevant settings.

Also fix some other warnings while I'm here.