emit @implements on 'implements' of type+value
If tsickle gets input that contains a type-value conflict, e.g.
  interface X { ... }
  const X = 3;
then it drops the 'interface' declaration in the output because
at the Closure level, the definition of "X" clobbers the @record.

Because of this, if someone later attempts to *use* the interface, e.g.
  class C implements X {}
tsickle would silently drop the '@implements' of that interface in
the Closure output.

But dropping @implements is dangerous in the presence of typed
optimzations; and this actually caused a real misoptimization in a
Google app.

The reason was that you actually can have legitimate type+value
interfaces in Closure:
  /** @interface */ class I { ... }
  I.someValue = 3;
and Clutz then dutifully converts this to the matching TypeScript,
and then if someone implements that interface, dropping the @implements
can cause a misoptimization.

In #1072 I convinced myself that we should instead never drop any
'implements' clause.  If the user writes the above pattern with X
and C, we should just compile fail instead, because there is no
way we can safely emit it.

In the test case here I added some usage of 'export import', but it
turns out that was irrelevant to this bug; it's just that when Clutz
started using 'export import' it got better at correctly describing
the above 'class I' pattern, which cascaded into this tsickle failure.
The clutz 'export import' change is
  https://github.com/angular/clutz/commit/cc9c9e6c8ca704aa05b5a3ec6e15ddf73d0c9a23