use mangled name for refs to type/value types
When there's a type/value conflict, we emit a mangled name for the type.
This change makes tsickle emit references to those mangled names when
the types are used.

Importantly, we need to be careful to only use mangled names in the case
where we're certain that the JS underneath was emitted via tsickle
already.  To do so we check the header that we added to the d.ts emit.