print a header on emitted d.ts
An upcoming change makes tsickle need to know whether a given module
was processed by tsickle or not.  This change makes tsickle emit a
constant string at the head of each d.ts file that it can look for.

For example, when tsickle compiles library a.ts, it emits a.d.ts.
Then when compiling b.ts that imports from a, it still knows (from
the tag in a.d.ts) that the JS emitted for a has been processed.

This change adds the header but does not make use of it.