WIP. Adds failing tests for @this {*} annotations.
This is an attempt to roll forward #1064 by just adds the requirement
through the goldens.

The following modifications need to be added to the @this {*} emitter:

- static async methods need to emit `@this {*}`
- arrow functions cannot have @this. It appears closure seems to ignore it,
likely because this is bound in arrow functions so it makes no sense to
have that annotation. Instead, this PR proses we emit @this {*} in the
enclosing `function` for top-level functions. Which should cover a lot
of cases, but not all.
- this golden also ensure that we do not emit `@this` for fields
assigned directly with an async function. It appears that in these cases
Closure uses type-inference on this and emitting @this, is detrimental
because it triggers a possible violation conformance check. This was why