always emit clutz types in heritage clauses
When there is a type-value conflict, e.g.
  interface B {}
  const B = 3;
then tsickle doesn't emit a Closure type for the interface.

Then when
  class A implements B
tsickle omits the @implements for B, because B doesn't exist.

But when B comes from a Clutz definition, we know that the type
B does exist on the Closure side, and we must emit the @implements.

This change keys off a Clutz-emitted file header to control this.

Partial fix of #1072.