fix(perf): cache symbol strings within modules.
tsickle has bad performance for certain specific patterns, union types
with many members that are declared in namespaces with many members.

TypeScript compiler expands unions eagerly, including literal type
members. E.g. `undefined|SomeEnum` gets represented as
`undefined|SomeEnum.A|SomeEnum.B|...`. We then generate the type for
each member and deduplicate. For large enums (in the wild we see >2000
members), this means we generate many symbolToString calls.

That wouldn't be so bad, but it turns out `symbolToString`, or rather
the TS API `symbolToEntityName` that it calls, can be very slow.

Consider this example:

    namespace foo {
      enum OurEnum {A, B, C, ...}
    }
    namespace foo {
      // more declarations
    }
    namespace foo {
      // even more declarations
    }
    ...

`symbolToEntityName` is called for each enum member, then gets the
parent symbol chain for the enum member, i.e. `[foo, OurEnum]` here, and
then for each declaration of `foo` inspects whether that particular
declaration is better to name the symbol (e.g. would have a more
accessible file name/path for the user, be shorter in the current scope,
etc.).

This looks esoteric, but it turns out this is exactly the kind of code
that Clutz generates for top level namespaces with many declarations in
separate files contained in them. We counted >100'000 declarations for a
specific root namespace, with enums with many members, causing multi
minute tsickle runtimes on very small sources.

The cache is a quick fix to avoid the duplicated work. It might be
possible to (a) avoid generating these in Clutz and (b) improve
performance of `symbolToEntityName`.