Variables first translate the type, then the body.
This fixes the test failure where we first translated the body, which
created a cache entry for the `T` symbol, and only then blacklisted the
symbol when generating the outer type.

This still seems not entirely right - the same symbol might be
blacklisted in one context, but might not in another, so we should
consider adding and removing from the backlist. That's complicated by
(currently) just using a hash map, we need to be careful not to
overwrite previous entries.