add @this annotations where needed for async functions
This is our third attempt in this space, with a totally new strategy.

Background: Closure is unhappy with the TS emit of async/await, because
it blindly refers to this.  Our previous attempts tried to add @this
tags where needed but ended up missing cases covering too much.

This new approach instead relies on Closure to generally infer this
appropriately.  The only case where we have a real problem is when
a non-arrow function uses 'this', because then 'this' really doesn't
refer to anything.  So this patch just adds @this to those functions.