Update docs on developing tsickle.
There's an annoying catch-22 between using `bazel run @nodejs//:yarn` to
install dependencies and using the bazel version installed by yarn,
which would require using `yarn install` to get bazel in the first
place.

In practice, I've had fewer problems with the bazel version, so this
changes the docs to recommend installing a local bazel version and using
it to install dependencies. This also removes the workaround around
`bazel run` and the associated script; I believe this has been fixed a
long time ago in bazel itself.

This change also deletes the .vscode/launch configuration and changes
documentation to warn on debugging using VS Code. I think this has been
broken ever since we moved to bazel and it seems impossible to get VS
Code to understand the mapping between paths in bazel-bin and the
original source tree.